+++ launcher/cli/apps.go
@@ -47,6 +47,7 @@ import (
 	"github.com/dfuse-io/dstore"
 	nodeosManagerApp "github.com/dfuse-io/manageos/app/nodeos_manager"
 	nodeosMindreaderApp "github.com/dfuse-io/manageos/app/nodeos_mindreader"
+	nodeosMindreaderStdinApp "github.com/dfuse-io/manageos/app/nodeos_mindreader_stdin"
 	"github.com/dfuse-io/manageos/mindreader"
 	mergerApp "github.com/dfuse-io/merger/app/merger"
 	pbblockmeta "github.com/dfuse-io/pbgo/dfuse/blockmeta/v1"
@@ -192,6 +193,50 @@ func init() {
 
 			// Can we detect a nil interface
 			return nil, nil
+		},
+	})
+	launcher.RegisterApp(&launcher.AppDef{
+		ID:          "mindreader-stdin",
+		Title:       "deep-mind reader from stdin",
+		Description: "deep-mind reader from stdin, does not start nodeos itself",
+		MetricsID:   "mindreader-stdin",
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/manageos/(app/nodeos_mindreader_stdin|mindreader).*", []zapcore.Level{zap.WarnLevel, zap.WarnLevel, zap.InfoLevel, zap.DebugLevel}),
+		RegisterFlags: func(cmd *cobra.Command) error {
+			return nil
+		},
+		FactoryFunc: func(modules *launcher.RuntimeModules) (launcher.App, error) {
+			dfuseDataDir, err := dfuseAbsoluteDataDir()
+			if err != nil {
+				return nil, err
+			}
+			archiveStoreURL := mustReplaceDataDir(dfuseDataDir, viper.GetString("common-oneblock-store-url"))
+			mergeArchiveStoreURL := mustReplaceDataDir(dfuseDataDir, viper.GetString("common-blocks-store-url"))
+
+			consoleReaderFactory := func(reader io.Reader) (mindreader.ConsolerReader, error) {
+				return codec.NewConsoleReader(reader)
+			}
+			//
+			consoleReaderBlockTransformer := func(obj interface{}) (*bstream.Block, error) {
+				blk, ok := obj.(*pbcodec.Block)
+				if !ok {
+					return nil, fmt.Errorf("expected *pbcodec.Block, got %T", obj)
+				}
+
+				return codec.BlockFromProto(blk)
+			}
+
+			return nodeosMindreaderStdinApp.New(&nodeosMindreaderStdinApp.Config{
+				ArchiveStoreURL:            archiveStoreURL,
+				MergeArchiveStoreURL:       mergeArchiveStoreURL,
+				MergeUploadDirectly:        viper.GetBool("mindreader-merge-and-store-directly"),
+				GRPCAddr:                   viper.GetString("mindreader-grpc-listen-addr"),
+				MindReadBlocksChanCapacity: viper.GetInt("mindreader-blocks-chan-capacity"),
+				WorkingDir:                 mustReplaceDataDir(dfuseDataDir, viper.GetString("mindreader-working-dir")),
+				DisableProfiler:            viper.GetBool("mindreader-disable-profiler"),
+			}, &nodeosMindreaderStdinApp.Modules{
+				ConsoleReaderFactory:     consoleReaderFactory,
+				ConsoleReaderTransformer: consoleReaderBlockTransformer,
+			}), nil
 		},
 	})